refactor: create a new quotation form component (only for desktop)
